array addition

+++ DSA/bubble_sort.cpp
@@ -3,29 +3,29 @@ using namespace std;
 
 int main()
 {
-    int n;
+    long n;
     cin >> n;
-    int arr[n];
-    for (int i = 0; i < n; i++)
+    long arr[n];
+    for (long i = 0; i < n; i++)
     {
         cin >> arr[i];
     }
-
-    int counter = 0;
-    while (counter < n - 1)
+// Bubble Sort
+    long counter = 0;
+    while (counter < n-1)
     {
-        for (int i = 0; i < n - counter; i++)
+        for (long i = 0; i<n-counter; i++)
         {
             if (arr[i] > arr[i + 1])
             {
-                int temp = arr[i];
+                long temp = arr[i];
                 arr[i] = arr[i + 1];
                 arr[i + 1] = temp;
             }
         }
         counter++;
     }
-    for (int i = 0; i < n; i++)
+    for (long i = 0; i < n; i++)
     {
         cout << arr[i] << "|";
     }
